refactor: organiza melhor a estrutura dos arquivos das páginas
