+++ src/components/common/index.ts
@@ -1,0 +1,2 @@
+// Componentes comuns/reutilizáveis
+export { default as ComingSoon } from './ComingSoon';
