+++ src/components/icons/index.ts
@@ -1,0 +1,2 @@
+// Ícones
+export { default as SearchIcon } from './SearchIcon';
