+++ src/components/index.ts
@@ -1,20 +1,5 @@
-// Componentes principais
-export { default as Header } from './Header';
-export { default as Footer } from './Footer';
-export { default as Home } from './Home';
-export { default as WelcomeSection } from './WelcomeSection';
-export { default as QuickActions } from './QuickActions';
-export { default as EmprestimosTable } from './EmprestimosTable';
-
-// Páginas
-export { default as Eventos } from './Eventos';
-export { default as Jogos } from './Jogos';
-export { default as Participantes } from './Participantes';
-export { default as Emprestimos } from './Emprestimos';
-export { default as Login } from './Login';
-
-// Componentes comuns
-export { default as ComingSoon } from './common/ComingSoon';
-
-// Ícones
-export { default as SearchIcon } from './icons/SearchIcon';
+// Re-exports organizados por categoria
+export * from './layout';
+export * from './ui';
+export * from './common';
+export * from './icons';
