+++ src/components/layout/index.ts
@@ -1,0 +1,3 @@
+// Componentes de layout
+export { default as Header } from './Header';
+export { default as Footer } from './Footer';
