+++ src/components/ui/WelcomeSection.tsx
