+++ src/components/ui/index.ts
@@ -1,0 +1,4 @@
+// Componentes de interface/UI
+export { default as WelcomeSection } from './WelcomeSection';
+export { default as QuickActions } from './QuickActions';
+export { default as EmprestimosTable } from './EmprestimosTable';
