+++ src/pages/Participantes.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import ComingSoon from "./common/ComingSoon";
+import { ComingSoon } from "../components";
 
 const Participantes: React.FC = () => {
     return <ComingSoon pageName="Participantes" />;
