+++ src/pages/index.ts
@@ -1,0 +1,7 @@
+// Páginas da aplicação
+export { default as Home } from './Home';
+export { default as Eventos } from './Eventos';
+export { default as Jogos } from './Jogos';
+export { default as Participantes } from './Participantes';
+export { default as Emprestimos } from './Emprestimos';
+export { default as Login } from './Login';
